BREAKING CHANGE: expect manual save_state! in init
After MGMT and RPT, there's been a transition away from
a state singleton and towards individual calls to
the (ns_)save/load functions from scrt-storage.
However the init implementation in scrt-contract
did not allow for deps to be mutably passed around.
This was because the macro rendered the init body
as a block inside a `get_store_rw.save` call.

This is now changed to match the behavior of tx handlers
which define a local save_state! macro that needs to be
called explicitly in order to update the contract's global
state.