feat(cli): implement "fadroma build" command
Takes a path to a ES module which exports-default
a `Record<string, ()=>Source[]>`, and the name of
the key in that Record corresponding to the list of
sources to build.

This is the first step to optimizing the dockerized build:
now that downstream Build.ts.md is declarative rather than
imperative, contracts from the same commit can be built
sequentially in the same container with the same cache.