fix(client): and types again + bump tag
It's like TypeScript is trying to make it more difficult to publish
robust software on purpose (knowing Microsoft's track record, yeah
it probably is.)

Thankfully I haven't published the package yet. Otherwise I'd be
wondering whether this commit goes in version 2.0.2 or 4.0.0.
What's the SemVer policy around changes to the type signatures?

Not only do you have to worry about changes to your code breaking
functionality downstream - now you also have to worry about changes to
your typing breaking whatever combination of typechecker/linter/etc
downstream uses!

Of course, Microsoft wouldn't be a 2-trillion USD company without being
realistic about the situation, and the reality is simple: downstream is
most likely to be largely ignoring the type checks. (As do we, as do the
frontend team.)

It's like they've decided to embrace and encourage the brokenness of
the JavaScript ecosystem under the guise of fixing it. Coincidence?
I doubt it. Using a type system that is far from expressive enough to
cover the flexibility of JavaScript, TypeScript provides strong
appearances of correctness, at the same time discouraging developers
from verifying the *actual* correctness of their code (by writing tests)
because "the types check out so it's fine".

And, as noticed above, most of the time the types don't even check out -
Developers are led to trust the instant feedback of the IDE, instead of
verifying at the actual behavior of the program. This is bad.

/rant