release(client-scrt@2.0.0): type fixes and version bumps
It seems that //@ts-ignore can only ignore the whole line and
you can't make it ignore the specific token that is causing the
type error.

I may be wrong - can't be bothered to test this extensively.
But if so, it means every //@ts-ignore that you add to bypass
the inexpressiveness of the type system makes the whole thing
more error-prone (because there could be other errors on that
line, and, what's worse, they could spontaneously appear as
results of changes in completely different places of the codebase.)

So it's not a true escape hatch, more like a booby trap.