fix(snip20): fix type of unreachable condition
This series of commits showed me one thing:
TS type inference leaves much to be desired.

One would expect them to have gotten it right
by now - they are well funded, and are imposing
an arbitrary subset of their own design on top
of the language.

But apparently, this shit is hard.
Who would've known, right?