refactor: reduce number of packages
`@fadroma/core` becomes `@fadroma/agent` and is officially
named Fadroma Agent, to correspond with its role of scriptable
"user agent". This is the base package for writing client libraries
(although currently it is consumed through `@fadroma/scrt`, which
provides the Secret Network implementation).

`@fadroma/mocknet` is now merged into `@fadroma/agent`,
bundling mocknet functionality alongside the isomorphic core.
This enables production builds of WASM contracts that conform
to the CosmWasm API to be executed in a simulated environment
based on the default WebAssembly runtime.

`@fadroma/ops` now combines `@fadroma/build`, `@fadroma/deploy`,
`@fadroma/devnet`, and `@fadroma/project`. This package contains
the functionality which depends on having a POSIX environment
(filesystem, sockets for talking to Docker, "shelling out" to
native executables, etc.)

The changes to directory and package hierarchies do not imply
changes to the class hierarchy, but reduces the number of individual
packages to be maintained on NPM. Upon release of Fadroma 1.0.0,
the numerous legacy packages on NPM will be marked as deprecated,
and/or unpublished.