test+feat(agent): define 1 new error and 3 new warnings
and comment out some tests that need to be reviewed (updated/removed).

next: implementing the isomorphic FetchUploader in @fadroma/agent.
now that Node has built-in fetch that works equally well with
file:/// urls (right? right? 🤞); also since the core deploy logic
is in the isomorphic module (@fadroma/agent) but the uploader is
in the posix-bound one (@fadroma/ops), providing the uploader
asap (~raii) can be made less of a pain by this.