fix(devnet): update ready message
scrt 1.8 had changed the format of their logs,
so instead of 'indexed block' we need to wait
for 'Done verifying block height'.

also gave that part of the system
(waiting for logs of docker containers)
a slight overhaul. could be sturdier and
less verbose, as sync with docker is imperfect
(sometimes a devnet container is left running).