fix+test(ops): lolwut all tests pass
73% coverage is a bit low - but enough to wrap up this round of
patching, and start looking at things through the doc lens again.

for now, the doctests confirm that that the pieces work when taken
separately, and their apis are about as hassle-free as the underlying
implementation currently allows.

devnet tests run quite quickly and exit cleanly, so what remains
is basically one big integration test that confirms that the user
can create a project, build it, deploy it to devnet, and export
the devnet image, all using the default workflow (no scripting,
only CLI - Deployment class is generated by project wizard); and
then (a) export a client lib and (b) run scripts in dev context.