fix(beta.66): package publishing still perilous
I think the root of the issue is because PNPM doesn't support
resolving versions that differ in the .rc.X segment
(which is legit semver from the spec). So setting "workspace:^1"
doesn't pick up "^1.0.0.rc-3" and publishing interdependent packages
is especially painful in this prerelease phase where it is done most
frequently. ouch!