perf(upload,mocknet): special case
don't force reupload to mocknet if the code is already there
mocknet is not stateful in the sense that the receipts written
become invalid the moment the mocknet gets garbage collected
(or the process ends). still, we can check if some code is
already present in the current mocknet incarnation.

it is also quite easy to save the state of the mocknet
(perhaps in a mocknet.json like the devnet.json) but
that exercise is left for another time